tqblogscn/requirements.txt
amqp == 2.2.2 
APScheduler == 3.5.1 
asn1crypto == 0.24.0 
attrs == 17.3.0 
Automat == 0.6.0 
billiard == 3.5.0.3 
celery == 4.1.0 
certifi == 2017.11.5 
cffi == 1.11.2 
chardet == 3.0.4 
click == 6.7 
constantly == 15.1.0 
cryptography == 2.1.4 
cssselect == 1.0.1 
diff-match-patch == 20121119 
Django == 1.10 
django-baseurl == 1.1 
django-crispy-forms == 1.7.0 
django-filter == 1.1.0 
django-formtools == 2.1 
django-future == 0.2.3 
django-haystack == 2.6.1 
django-import-export == 0.6.1 
django-picklefield == 1.0.0 
django-redis == 4.8.0 
django-reversion == 2.0.12 
django-tinymce == 2.6.0 
djangorestframework == 3.7.7 
et-xmlfile == 1.0.1 
Flask == 0.12.2 
future == 0.16.0 
gunicorn == 19.7.1 
html5lib == 1.0.1 
httplib2 == 0.9.2 
hyperlink == 17.3.1 
idna == 2.6 
incremental == 17.5.0 
itsdangerous == 0.24 
jdcal == 1.3 
jieba == 0.39 
Jinja2 == 2.10 
jsmin == 2.2.2 
kombu == 4.1.0 
lxml == 4.1.1 
Markdown == 2.6.11 
MarkupSafe == 1.0 
odfpy == 1.3.6 
olefile == 0.44 
openpyxl == 2.4.9 
parsel == 1.2.0 
Pillow == 4.3.0 
pip == 9.0.2 
pyasn1 == 0.4.2 
pyasn1-modules == 0.2.1 
pycparser == 2.18 
pycryptodome == 3.4.7 
PyDispatcher == 2.0.5 
Pygments == 2.2.0 
pymongo == 3.6.1 
PyMySQL == 0.7.11 
pyOpenSSL == 17.5.0 
python-alipay-sdk == 1.4.0 
pytz == 2017.2 
PyYAML == 3.12 
queuelib == 1.4.2 
redis == 2.10.6 
requests == 2.18.4 
Scrapy == 1.4.0 
scrapy-redis == 0.6.8 
selenium == 3.11.0 
service-identity == 17.0.0 
setuptools == 38.2.4 
six == 1.11.0 
tablib == 0.12.1 
tornado == 4.5.2 
Twisted == 17.9.0 
tzlocal == 1.5.1 
unicodecsv == 0.14.1 
urllib3 == 1.22 
uWSGI == 2.0.15 
vine == 1.1.4 
w3lib == 1.18.0 
webencodings == 0.5.1 
Werkzeug == 0.13 
wheel == 0.30.0 
Whoosh == 2.7.4 
xlrd == 1.1.0 
xlwt == 1.3.0 
zope.interface == 4.4.3 
